move io GUI to gui module
